feat: Add an article about the birth of the website.
An article about the birth of the website has been added in order to let users understand the process of the birth of the website of Guangdong Gaozhou Middle School Computer Association. The article is co-written by Scattered Leaves and WesKwong, and describes in detail the background and development process of the project.